frontend commit

+++ backend/controllers/goalController.js
@@ -2,9 +2,11 @@ const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalModel");
 
+const User = require("../models/userModel");
+
 // Get goals, GET api/goals, Private access
 const getGoals = asyncHandler(async (req, res) => {
-  const goal = await Goal.find();
+  const goal = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goal);
 });
@@ -18,7 +20,9 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
+
   res.status(200).json(goal);
 });
 
@@ -31,9 +35,25 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found.");
   }
 
+  const user = await findById(req.user.id);
+
+  // check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
+  // update goal with the given id, returns object with updated text
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
+
   res.status(200).json(updatedGoal);
 });
 
@@ -45,6 +65,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found.");
   }
+
+  const user = await findById(req.user.id);
+
+  // check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   // deletes the goal with the given id, will return object with the deleted id
   await Goal.findByIdAndDelete(req.params.id);
   res.status(200).json({ id: req.params.id });
